refactor: separate page content into a `PageContent.vue` component
